requirements.txt
torch
torchvision
torchaudio
transformers
opencv-python
Pillow
timm
vit-pytorch
numpy
matplotlib
scipy
tqdm
pytest
jupyter
ipython
lmdb